MicrosoftOffice/requirements.txt
lxml==5.3.0
pillow==11.0.0
python-docx==1.1.2
python-pptx==1.0.2
typing_extensions==4.12.2
XlsxWriter==3.2.0
lxml==5.3.0
pillow==11.0.0
python-docx==1.1.2
python-pptx==1.0.2
typing_extensions==4.12.2
XlsxWriter==3.2.0
